+++ src/app/api-services/index.ts
@@ -1,7 +1,5 @@
-export * from './auth.service';
 export * from './models';
 export * from './api-service.module';
-export * from './preset.service';
 export * from './ro.service';
 export * from './summary.service';
 export * from './divine-pride.service';