+++ src/app/app-errors/unauthorized.ts
@@ -1,7 +1,0 @@
-import { AppErrorBase } from './base-error';
-
-export class Unauthorized extends AppErrorBase {
-  constructor() {
-    super('Unauthorized');
-  }
-}