+++ src/app/app.module.ts
@@ -6,7 +6,6 @@ import { AppLayoutModule } from './layout/app.layout.module';
 import { RoService } from './api-services/ro.service';
 import { PrettyJsonPipe } from './layout/prettier-json.pipe';
 import { ApiServiceModule } from './api-services';
-import { JwtModule } from '@auth0/angular-jwt';
 import { SummaryService } from './api-services/summary.service';
 
 const customComponent = [PrettyJsonPipe];
@@ -17,11 +16,6 @@ const customComponent = [PrettyJsonPipe];
     AppRoutingModule,
     AppLayoutModule,
     ApiServiceModule,
-    JwtModule.forRoot({
-      config: {
-        tokenGetter: () => localStorage.getItem('accessToken'),
-      },
-    }),
   ],
   providers: [{ provide: LocationStrategy, useClass: HashLocationStrategy }, RoService, SummaryService, ...customComponent],
   bootstrap: [AppComponent],